chams update
Added obs bypassed chams (players, entitys, hands)
Added more safety for lua dumper

fix some bugs
fix player box
code rewrite:
added the ability to create custom materials
added key values
fixed get_entity_from_handle

+++ sdk/source-sdk/math/vector.h
@@ -6,7 +6,7 @@ using matrix4x4 = float[16];
 
 struct box_t
 {
-	int left, top, right, bottom;
+	float left, top, right, bottom;
 };
 
 class c_vector
